+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 0.0.6 (2020-06-19)
+
+### Bug Fixes
+
+* eslint编译未通过 ([d636c77](https://github.com/Iecy/eagle/commit/d636c77aa672500ff8e84ea7ba4f63cc66841a6c))
+* 测试覆盖率 ([35d2767](https://github.com/Iecy/eagle/commit/35d2767c0da97c0f60308c943052df0ed2b5c669))
+* 测试覆盖率 ([2603a03](https://github.com/Iecy/eagle/commit/2603a03951c6e6be7034f80bf39f5f308aaaf129))
+
+
+### Features
+
+* drop/dropLast/take/slice/numberFormat ([0f96e61](https://github.com/Iecy/eagle/commit/0f96e616440b1e7e16e4fbb2efedf26f5abf6d10))
+* init ([a392fb4](https://github.com/Iecy/eagle/commit/a392fb412820db33ef9dd59d761afec7c8cf1170))
+* 增加isEmpty/equals/keys方法 ([3e5b72a](https://github.com/Iecy/eagle/commit/3e5b72a9b19c6fd02250b8654c0ec505a80d8149))
+* 增加trim ([d8d8323](https://github.com/Iecy/eagle/commit/d8d8323bfdec74cff6fb05bb123fe0416aed3f34))
+* 增加电话/身份证号码的验证 ([735d9a2](https://github.com/Iecy/eagle/commit/735d9a2f5a87c360a039b25936e548b5f540ec55))
+* 新增empty.clone ([71e761d](https://github.com/Iecy/eagle/commit/71e761df35b0e22fdae55439a29dda4e2a9c319b))
+
 ## 0.0.2 (2020-04-14)